nlpoison/requirements.txt
transformers==4.2.1
torch==1.7.1
# torch==1.7.1+cu101
torchvision==0.8.2
# torchvision==0.8.2+cu101 
# torchaudio==0.7.2 #-f https://download.pytorch.org/whl/torch_stable.html
pandas
pyyaml
wandb
sklearn
matplotlib
seaborn 
nltk
adversarial_robustness_toolbox==1.6.1
